casted x and y to float in relative coords
мб это фиксанет проблемы графика и то что relative корды всегда целочисленные

+++ src/main/java/the/kis/devs/g2i/Main.java
@@ -114,8 +114,8 @@ public class Main {
 
                 {
                     if(x > axisOffset + axisWidth && y < height - axisOffset - axisWidth) {
-                        float relativeX = round((x - axisOffset - axisWidth) / coordinateSystemSingleSection);
-                        float relativeY = round((coordinateSystemYLength - (y - axisOffset)) / coordinateSystemSingleSection);
+                        float relativeX = round(((float) x - axisOffset - axisWidth) / coordinateSystemSingleSection);
+                        float relativeY = round((coordinateSystemYLength - ((float) y - axisOffset)) / coordinateSystemSingleSection);
 
                         float linearFunctionRelativeY = doLinearFunction(relativeX, 0.5f, -3);
